# docs/swagger.yaml
basePath: /
definitions:
...
  models.Teacher:
    properties:
      class:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      subject:
        type: string
    type: object
host: localhost:443
info:
  contact: {}
  description: This is a REST API for managing Teachers, Students and Executives
  title: REST API Documentation
  version: "1.0"
paths:
...
  /allteachers/:
    delete:
      description: Delete all teachers from the system
      responses:
        "200":
          description: All teachers deleted successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Delete all teachers
      tags:
      - Teachers
...
  /teachers/:
    delete:
      consumes:
      - application/json
      description: Delete multiple teachers by their IDs
      parameters:
      - description: Teacher IDs to delete
        in: body
        name: ids
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: Teachers deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Delete multiple teachers
      tags:
      - Teachers
    get:
      description: Get a list of all teachers
      produces:
      - application/json
      responses:
        "200":
          description: List of teachers
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get all teachers
      tags:
      - Teachers
    patch:
      consumes:
      - application/json
      - application/json
      description: |-
        Partially update multiple teachers
        Partially update a specific teacher by ID
      parameters:
      - description: Teacher updates
        in: body
        name: updates
        required: true
        schema:
          items:
            additionalProperties: true
            type: object
          type: array
      - description: Teacher updates
        in: body
        name: updates
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: Teacher patched successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer// PatchTeacherHandler patches a specific teacher: []
      - Bearer: []
      summary: Patch teacher
      tags:
      - Teachers
      - Teachers
    post:
      consumes:
      - application/json
      description: Create one or multiple new teachers
      parameters:
      - description: Teacher data
        in: body
        name: teachers
        required: true
        schema:
          items:
            $ref: '#/definitions/models.Teacher'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Teachers created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Add new teachers
      tags:
      - Teachers
  /teachers/{id}:
    delete:
      description: Delete a specific teacher by their ID
      parameters:
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Teacher deleted successfully
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Delete teacher
      tags:
      - Teachers
    get:
      description: Get a specific teacher by their ID
      parameters:
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Teacher'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get teacher by ID
      tags:
      - Teachers
    patch:
      consumes:
      - application/json
      - application/json
      description: |-
        Partially update multiple teachers
        Partially update a specific teacher by ID
      parameters:
      - description: Teacher updates
        in: body
        name: updates
        required: true
        schema:
          items:
            additionalProperties: true
            type: object
          type: array
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: integer
      - description: Teacher updates
        in: body
        name: updates
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: Teacher patched successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer// PatchTeacherHandler patches a specific teacher: []
      - Bearer: []
      summary: Patch teacher
      tags:
      - Teachers
      - Teachers
    put:
      consumes:
      - application/json
      description: Update a specific teacher by their ID
      parameters:
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: integer
      - description: Teacher data to update
        in: body
        name: teacher
        required: true
        schema:
          $ref: '#/definitions/models.Teacher'
      produces:
      - application/json
      responses:
        "200":
          description: Teacher updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update teacher
      tags:
      - Teachers
  /teachers/{id}/studentcount/:
    get:
      description: Get the number of students taught by a specific teacher
      parameters:
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Student count
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get student count for teacher
      tags:
      - Teachers
      - Students
  /teachers/{id}/students/:
    get:
      description: Get a list of all students taught by a specific teacher
      parameters:
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of students
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get students by teacher
      tags:
      - Teachers
      - Students
schemes:
- https
securityDefinitions:
  Bearer:
    description: JWT Bearer token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"